Archive/pyproject.toml
[project]
name = "Copilot-Plugins-Azure-Functions-deployment-v2"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Sumsam Ali", email = "sumsam.ali@imperiumdynamics.com"},
]
dependencies = [
    "azure-functions",
    "requests",
    "azure-identity",
    "python-dateutil",
    "pytz",
    "validate-email-address",
    "fuzzywuzzy",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
